+++ docs/CHANGELOG.md
@@ -7,7 +7,8 @@
 4. Upgrade discord.js v12 -> v13
 5. Add /help bounty, fix grammar for text, simplify bot commands
 6. Capture bounty description and criteria from bot interaction messages
-7. Add POAP tracking command for CC call
+7. Allow publication of the same bounty multiple times for level 3+ users
+8. Add POAP tracking command for CC call
 
 ## 1.2.1-RELEASE (2021-08-12)